+++ packages/yak_result/lib/src/classes/result.dart
@@ -1,21 +1,10 @@
 import 'package:meta/meta.dart';
 import 'package:yak_utils/yak_utils.dart';
-import 'all.dart';
 
-/// represent the outcome of a functionf
+/// represent the outcome of a function
 abstract class Result<T> {
   /// allows a const constructor
   const Result();
-
-  const factory Result.success(T data) = Success;
-  const factory Result.pending() = Pending;
-  factory Result.failure({Object? reason, StackTrace? stackTrace}) = Failure;
-
-  @override
-  bool operator ==(Object other) => hashCode == other.hashCode;
-
-  @override
-  int get hashCode => Object.hash(runtimeType, T.runtimeType);
 
   @override
   @nonVirtual
@@ -23,4 +12,14 @@ abstract class Result<T> {
 
   /// allow compatibility with [dart:convert]
   Json toJson();
+}
+
+/// a result returning value
+abstract class ValueResult<T> extends Result<T> {
+  const ValueResult();
+}
+
+/// a result without returning value
+abstract class VoidResult<T> extends Result<T> {
+  const VoidResult();
 }