+++ packages/yak_result/lib/src/typedef/future_result.dart
@@ -1,0 +1,12 @@
+import 'dart:async';
+
+import '../all.dart';
+
+/// a [FutureOr] [Result]
+typedef FutureResult<T> = FutureOr<Result<T>>;
+
+/// a [FutureOr] [ValueResult]
+typedef FutureValueResult<T> = FutureOr<ValueResult<T>>;
+
+/// a [FutureOr] [VoidResult]
+typedef FutureVoidResult = FutureOr<VoidResult<void>>;