+++ packages/yak_result/lib/src/typedef/nullary_function.dart
@@ -1,9 +1,16 @@
-import 'package:yak_utils/yak_utils.dart' show Nullary, NullaryAsync;
+import 'dart:async';
+import 'package:yak_utils/yak_utils.dart' show Nullary;
 
 import '../all.dart';
 
-/// an nullary function that returns a Result
-typedef ResultNullary<T> = Nullary<Result<T>>;
+/// an [Nullary] [Function] that returns a [Result]
+typedef ResultNullary<T> = Nullary<ValueResult<T>>;
 
-/// an nullary async function that returns a Result
-typedef ResultNullaryAsync<T> = NullaryAsync<Result<T>>;
+/// an [Nullary] [async] [Function] that returns a [Result]
+typedef ResultNullaryAsync<T> = Nullary<FutureOr<ValueResult<T>>>;
+
+/// an [Nullary] function that returns a [VoidResult]
+typedef VoidResultNullary = Nullary<VoidResult<void>>;
+
+/// an [Nullary] [async] [Function] that returns a [VoidResult]
+typedef VoidResultNullaryAsync = Nullary<FutureOr<VoidResult<void>>>;